requirements.txt
pandas
mplfinance
requests
plotly
dotenv
PyYAML